Add game file and spec, remove comments in runner file
